requirements.txt
packaging
transformer
datasets
tqdm
librosa
ninja
flash-attn --no-build-isolation
